Maven added. Folders structure rearranged for Maven
