+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/AdditionExpressionNode.java
@@ -22,63 +22,76 @@
  * THE SOFTWARE.
  */
 
-package uk.co.cogitolearning.cogpar;
+package co.uk.cogitolearning.cogpar;
 
 import java.util.Iterator;
 
 /**
- * An ExpressionNode that handles divisions. The node can hold
- * an two number of factors that are divided.
+ * An ExpressionNode that handles additions and subtractions. The node can hold
+ * an arbitrary number of terms that are either added or subtraced from the sum.
  * 
  */
-public class DivExpressionNode extends SequenceExpressionNode
+public class AdditionExpressionNode extends SequenceExpressionNode
 {
+
   /**
    * Default constructor.
    */
-  public DivExpressionNode()
+  public AdditionExpressionNode()
   {}
 
   /**
-   * Constructor to create a multiplication with the first term already added.
-   *
-   * @param a
+   * Constructor to create an addition with the first term already added.
+   * 
+   * @param node
    *          the term to be added
    * @param positive
-   *          a flag indicating whether the term is multiplied or divided
+   *          a flag indicating whether the term is added or subtracted
    */
-  public DivExpressionNode(ExpressionNode a, boolean positive)
+  public AdditionExpressionNode(ExpressionNode node, boolean positive)
   {
-    super(a, positive);
+    super(node, positive);
   }
 
   /**
-   * Returns the type of the node, in this case ExpressionNode.MULTIPLICATION_NODE
+   * Returns the type of the node, in this case ExpressionNode.ADDITION_NODE
    */
   public int getType()
   {
-    return ExpressionNode.DIVISION_NODE;
+    return ADDITION_NODE;
   }
 
+  /**
+   * Returns the value of the sub-expression that is rooted at this node.
+   * 
+   * All the terms are evaluated and added or subtracted from the total sum.
+   */
   public double getValue()
   {
-    // Not used
-    return 0;
+    double sum = 0.0;
+    for (Term t : terms)
+    {
+      if (t.positive)
+        sum += t.expression.getValue();
+      else
+        sum -= t.expression.getValue();
+    }
+    return sum;
   }
 
   /**
    * Implementation of the visitor design pattern.
    * 
    * Calls visit on the visitor and then passes the visitor on to the accept
-   * method of all the terms in the product.
+   * method of all the terms in the sum.
    * 
    * @param visitor
    *          the visitor
    */
   public void accept(ExpressionNodeVisitor visitor)
   {
-    visitor.visit(this);  
-    for (Term t: terms)
+    visitor.visit(this);
+    for (Term t : terms)
       t.expression.accept(visitor);
   }
