+++ CogPar/src/main/java/co/uk/cogitolearning/cogpar/CalculateValue.java
@@ -1,4 +1,4 @@
-package uk.co.cogitolearning.cogpar;
+package co.uk.cogitolearning.cogpar;
 
 import java.util.ArrayList;
 import java.util.Collections;
